need a better shred builder at this point
