+++ ledger/src/shred/shred_code.rs
@@ -28,8 +28,6 @@ impl ShredCode {
     dispatch!(fn coding_header(&self) -> &CodingShredHeader);
 
     dispatch!(pub(super) fn common_header(&self) -> &ShredCommonHeader);
-    #[cfg(feature = "dev-context-only-utils")]
-    dispatch!(pub(super) fn common_header_mut(&mut self) -> &mut ShredCommonHeader);
     dispatch!(pub(super) fn erasure_shard(&self) -> Result<&[u8], Error>);
     dispatch!(pub(super) fn erasure_shard_index(&self) -> Result<usize, Error>);
     dispatch!(pub(super) fn first_coding_index(&self) -> Option<u32>);
