+++ ledger/src/shred/shred_data.rs
@@ -21,8 +21,6 @@ impl ShredData {
     dispatch!(fn data_header(&self) -> &DataShredHeader);
 
     dispatch!(pub(super) fn common_header(&self) -> &ShredCommonHeader);
-    #[cfg(feature = "dev-context-only-utils")]
-    dispatch!(pub(super) fn common_header_mut(&mut self) -> &mut ShredCommonHeader);
     dispatch!(pub(super) fn erasure_shard(&self) -> Result<&[u8], Error>);
     dispatch!(pub(super) fn erasure_shard_index(&self) -> Result<usize, Error>);
     dispatch!(pub(super) fn into_payload(self) -> Payload);
@@ -62,7 +60,7 @@ impl ShredData {
         version: u16,
         fec_set_index: u32,
     ) -> Self {
-        Self::from(legacy::ShredData::new_from_data(
+        Self::from(merkle::ShredData::new_from_data(
             slot,
             index,
             parent_offset,
