+++ turbine/src/broadcast_stage/fail_entry_verification_broadcast_run.rs
@@ -103,7 +103,6 @@ impl BroadcastRun for FailEntryVerificationBroadcastRun {
             Some(self.chained_merkle_root),
             self.next_shred_index,
             self.next_code_index,
-            true, // merkle_variant
             &self.reed_solomon_cache,
             &mut ProcessShredsStats::default(),
         );
@@ -123,7 +122,6 @@ impl BroadcastRun for FailEntryVerificationBroadcastRun {
                 Some(self.chained_merkle_root),
                 self.next_shred_index,
                 self.next_code_index,
-                true, // merkle_variant
                 &self.reed_solomon_cache,
                 &mut ProcessShredsStats::default(),
             );
@@ -137,7 +135,6 @@ impl BroadcastRun for FailEntryVerificationBroadcastRun {
                 Some(self.chained_merkle_root),
                 self.next_shred_index,
                 self.next_code_index,
-                true, // merkle_variant
                 &self.reed_solomon_cache,
                 &mut ProcessShredsStats::default(),
             );
